+++ smtp_by_api.go
@@ -1,9 +1,5 @@
 package emailverifier
 
-const (
-	YAHOO = "yahoo"
-)
-
 type smtpAPIVerifier interface {
 	// isSupported the specific host supports the check by api.
 	isSupported(host string) bool